+++ src/api/hub_handler.rs
@@ -1,8 +1,14 @@
+use std::fmt;
 use std::marker::PhantomData;
+
+use serde::de::DeserializeOwned;
 
 use crate::api::{ApiClient, Authenticated, Unauthenticated};
 use crate::error::Error;
-use crate::responses::{ChildDeviceListResult, HubDeviceInfoResult};
+use crate::requests::TapoRequest;
+use crate::responses::{
+    ChildDeviceListResult, ChildDeviceResult, HubDeviceInfoResult, TapoResponseExt,
+};
 
 /// Handler for the [H100](https://www.tapo.com/en/search/?q=H100) hubs.
 pub struct HubHandler<S = Unauthenticated> {
@@ -46,13 +52,34 @@ impl HubHandler<Authenticated> {
     /// Gets *child device list* as [`crate::responses::ChildDeviceListResult`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API
     /// or to support all the possible devices connected to the hub.
-    pub async fn get_child_device_list(&self) -> Result<ChildDeviceListResult, Error> {
-        self.client.get_child_device_list().await
+    pub async fn get_child_device_list(&self) -> Result<Vec<ChildDeviceResult>, Error> {
+        self.client
+            .get_child_device_list::<ChildDeviceListResult>()
+            .await
+            .map(|r| r.devices)
     }
 
     /// Gets *child device list* as [`serde_json::Value`].
     /// It contains all the properties returned from the Tapo API.
     pub async fn get_child_device_list_json(&self) -> Result<serde_json::Value, Error> {
         self.client.get_child_device_list().await
+    }
+
+    /// Gets *child device component list* as [`serde_json::Value`].
+    /// This information is useful in debugging or when investigating new functionality to add.
+    pub async fn get_child_device_component_list_json(&self) -> Result<serde_json::Value, Error> {
+        self.client.get_child_device_component_list().await
+    }
+
+    /// Internal method that's called by functions of the child devices.
+    pub(crate) async fn control_child<R>(
+        &self,
+        device_id: String,
+        request_data: TapoRequest,
+    ) -> Result<R, Error>
+    where
+        R: fmt::Debug + DeserializeOwned + TapoResponseExt,
+    {
+        self.client.control_child(device_id, request_data).await
     }
 }