+++ src/requests.rs
@@ -1,11 +1,14 @@
 //! Tapo request objects.
 
 mod color;
+mod control_child;
 mod energy_data_interval;
 mod get_energy_data;
+mod get_trigger_logs;
 mod handshake;
 mod lighting_effect;
 mod login_device;
+mod multiple_request;
 mod secure_passthrough;
 mod set_device_info;
 mod tapo_request;
@@ -15,8 +18,11 @@ pub use energy_data_interval::*;
 pub use lighting_effect::*;
 pub use set_device_info::*;
 
+pub(crate) use control_child::*;
 pub(crate) use get_energy_data::*;
+pub(crate) use get_trigger_logs::*;
 pub(crate) use handshake::*;
 pub(crate) use login_device::*;
+pub(crate) use multiple_request::*;
 pub(crate) use secure_passthrough::*;
 pub(crate) use tapo_request::*;