+++ src/responses.rs
@@ -1,6 +1,7 @@
 //! Tapo response objects.
 
 mod child_device_list_result;
+mod control_child_result;
 mod decodable_result_ext;
 mod device_info_result;
 mod device_usage_result;
@@ -10,13 +11,16 @@ mod handshake_result;
 mod tapo_response;
 mod tapo_result;
 mod token_result;
+mod trigger_logs_result;
 
 pub use child_device_list_result::*;
 pub use device_info_result::*;
 pub use device_usage_result::*;
 pub use energy_data_result::*;
 pub use energy_usage_result::*;
+pub use trigger_logs_result::*;
 
+pub(crate) use control_child_result::*;
 pub(crate) use decodable_result_ext::*;
 pub(crate) use handshake_result::*;
 pub(crate) use tapo_response::*;